Feature/46254 bug fix evidence summary note information not reflecting selected compliance

+++ src/EA.Weee.RequestHandlers/AatfReturn/Internal/IAatfDataAccess.cs
@@ -30,7 +30,7 @@
 
         Task<List<short>> GetComplianceYearsForAatfByAatfId(Guid aatfId);
 
-        Task<Guid> GetAatfByAatfIdAndComplianceYear(Guid aatfId, short complianceYear);
+        Task<Guid> GetAatfByAatfIdAndComplianceYear(Guid aatfId, int complianceYear);
 
         Task<bool> IsLatestAatf(Guid id, Guid aatfId);
     }
